+++ db/migrate/20170208170715_create_reviews.rb
@@ -3,7 +3,7 @@ class CreateReviews < ActiveRecord::Migration[5.0]
     create_table :reviews do |t|
       t.column :name, :string
       t.column :details, :string
-      t.column :rating, :integer
+      t.column :rating, :float
       t.column :work_id, :integer
 
       t.timestamps
